requirements.txt
fastapi
uvicorn
google-generativeai
python-dotenv
googlesearch-python
pydantic